+++ src/main/resources/application.yaml
@@ -4,6 +4,7 @@ no.nav.security.jwt:
       discoveryurl: ${azure.app.well.known.url}
       accepted_audience: ${azure.app.client.id}
       cookiename: ID_token
+      proxy-url: http://webproxy.nais:8088
     tokenx:
       discoveryurl: ${TOKEN_X_WELL_KNOWN_URL}
       accepted_audience: ${TOKEN_X_CLIENT_ID}
